Extract focused_edit_recovery.rs from turn.rs (parent #680)
Move the focused-edit recovery cluster out of turn.rs into a new
src/agent/loop_run/focused_edit_recovery.rs sibling module:

- focused_edit_guidance_note / _for_policy
- focused_edit_compact_anchor_note
- focused_edit_first_slice_note / _uses_exact_anchor
- focused_edit_second_slice_note
- focused_edit_exact_recovery_anchor / _compact_recovery_anchor
- focused_edit_history / _minimal_history / _exact_anchor_history
- latest_page_copy_block_from_read
- latest_page_intro_copy_line_from_read
- extract_page_copy_block_from_numbered_read
- extract_page_intro_paragraph_from_numbered_read
- extract_page_intro_copy_line_from_numbered_read
- extract_compact_edit_anchor_from_numbered_read
- strip_read_line_number_prefix
- format_numbered_read_block
- is_page_component_target

All items are pub(super), no facade re-export (DR3-001). turn.rs is the
only in-crate consumer; production entry points are re-imported into
turn.rs scope, with the test-only symbols gated by #[cfg(test)].

turn.rs: 27,394 -> 27,018 (-376 LOC)
Cumulative from 39,489: -12,471 LOC (-31.6%)

Verification:
- cargo fmt --check
- cargo clippy --lib --all-targets -- -D warnings
- cargo test --lib (3,114 passed)
